User Management App/requirements.txt
bcrypt==4.0.1
blinker==1.6.3
cachelib==0.10.2
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
et-xmlfile==1.1.0
flask==3.0.0
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
flask-session==0.5.0
flask-wtf==1.2.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
openpyxl==3.1.2
packaging==23.2
pymongo==4.5.0
python-dotenv==1.0.0
werkzeug==3.0.0
wtforms==3.1.0
zipp==3.17.0